- Prepare hilt injection with ksp - Created repositories and get articles from remote

+++ network/src/main/java/com/davidmerchan/network/model/Hits.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Hits(
     @SerializedName("story_id")
-    val storyId: Int,
+    val storyId: Long,
     @SerializedName("story_title")
     val storyTitle: String?,
     @SerializedName("story_url")
